feat(hub-common): created a utility for public array related exports
affects: @esri/hub-common
