+++ packages/common/src/utils/index.ts
@@ -32,4 +32,4 @@ export * from "./isComboboxItemSelected";
 export * from "./wait";
 export * from "./IUserHubSettings";
 export * from "./IUserSiteSettings";
-export * from "./_array";
+export * from "./array";
